Fix integration tests.
